move command to separate package
