+++ Plugin/src/main/java/plugin/CustomStructuresPlugin.java
@@ -8,6 +8,7 @@ import java.util.Set;
 
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.event.Event;
+import plugin.command.CustomStructuresCommand;
 import plugin.structure.Structure;
 
 public class CustomStructuresPlugin extends JavaPlugin {
